+++ src/superface/config/prepare-ast.ts
@@ -7,9 +7,12 @@ import {
 } from '@superfaceai/ast';
 import { IFileSystem, Profile } from '@superfaceai/one-sdk';
 import { parseMap, parseProfile, Source } from '@superfaceai/parser';
-import { dirname, resolve as resolvePath } from 'path';
+import createDebug from 'debug';
+import { resolve as resolvePath } from 'path';
 
 import { MapUndefinedError, ProfileUndefinedError } from '../../common/errors';
+
+const debugSetup = createDebug('superface:testing:setup');
 
 export async function getProfileAst(
   profile: Profile | string,
@@ -25,10 +28,11 @@ export async function getProfileAst(
       throw new ProfileUndefinedError(profile);
     }
 
-    const profilePath = resolvePath(
-      dirname(superJson.path),
-      profileSettings.file
-    );
+    const profilePath = resolvePath(superJson.path, profileSettings.file);
+
+    debugSetup('Found profile settings in super.json:', profileSettings);
+    debugSetup('Profile resolved path:', profilePath);
+
     const content = await fileSystem.readFile(profilePath);
 
     if (content.isErr()) {
@@ -37,6 +41,8 @@ export async function getProfileAst(
 
     // TODO: make parser optional
     if (profilePath.endsWith('.supr')) {
+      debugSetup('Trying to parse profile:', profilePath);
+
       return parseProfile(new Source(content.value, profilePath));
     } else {
       return assertProfileDocumentNode(JSON.parse(content.value));
@@ -57,10 +63,14 @@ export async function getMapAst(
     throw new MapUndefinedError(profileId, providerName);
   }
 
-  const mapPath = resolvePath(
-    dirname(superJson.path),
-    profileProviderSettings.file
+  const mapPath = resolvePath(superJson.path, profileProviderSettings.file);
+
+  debugSetup(
+    'Found profile provider settings in super.json:',
+    profileProviderSettings
   );
+  debugSetup('Map resolved path:', mapPath);
+
   const content = await fileSystem.readFile(mapPath);
 
   if (content.isErr()) {
@@ -69,6 +79,8 @@ export async function getMapAst(
 
   // TODO: make parser optional
   if (mapPath.endsWith('.suma')) {
+    debugSetup('Trying to parse map:', mapPath);
+
     return parseMap(new Source(content.value, mapPath));
   } else {
     return assertMapDocumentNode(JSON.parse(content.value));
